livekit_selective_conference/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
pydantic==2.7.1
livekit-api==0.8.2
